refactor: move code from `__init__`
